+++ src/consumer/component/sources/license.js
@@ -1,6 +1,6 @@
 /** @flow */
-import * as fs from 'fs';
-import * as path from 'path';
+import fs from 'fs-extra';
+import path from 'path';
 import Source from './source';
 import { LICENSE_FILENAME } from '../../../constants';
 
@@ -8,12 +8,7 @@ export default class License extends Source {
   write(bitPath: string, force?: boolean = true): Promise<any> {
     const filePath = path.join(bitPath, LICENSE_FILENAME);
     if (!force && fs.existsSync(filePath)) return Promise.resolve();
-    return new Promise((resolve, reject) =>
-      fs.writeFile(filePath, this.src, (err, res) => {
-        if (err) return reject(err);
-        return resolve(res);
-      })
-    );
+    return fs.writeFile(filePath, this.src);
   }
 
   serialize() {