+++ src/utils/fs-exists-sync.js
@@ -1,6 +1,0 @@
-/** @flow */
-import fs from 'fs';
-
-export default function existsSync(path: string): Boolean {
-  return fs.existsSync(path);
-}