+++ src/utils/is-dir.js
@@ -1,5 +1,5 @@
 /** @flow */
-import fs from 'fs';
+import fs from 'fs-extra';
 import GeneralError from '../error/general-error';
 
 export default function isDir(userPath: string): boolean {